refactor(hooks): use new lib api & types
