+++ src/hooks/useDeleteTask.ts
@@ -1,19 +1,17 @@
-import { useMutation, useQueryClient } from '@tanstack/react-query';
-import axios from 'axios';
+import api from "@/libs/api";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
 
-const useDeleteTask = (id: string) => {
-  const queryClient = useQueryClient();
+export default function useDeleteTask(id: string) {
+	const queryClient = useQueryClient();
 
-  const mutation = useMutation({
-    mutationFn: () => {
-      return axios.delete(`http://localhost:8000/tasks/${id}`);
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['tasks'] });
-    },
-  });
+	const mutation = useMutation({
+		mutationFn: () => {
+			return api.delete(`http://localhost:8000/tasks/${id}`);
+		},
+		onSuccess: () => {
+			queryClient.invalidateQueries({ queryKey: ["tasks"] });
+		},
+	});
 
-  return mutation;
-};
-
-export default useDeleteTask;
+	return mutation;
+}
