+++ src/hooks/useGetTaskCounts.ts
@@ -1,25 +1,24 @@
-import { useEffect } from 'react';
-import tableDataReducers from '../components/utils/tableDataReducers';
-import TaskCounts from '../types/TaskCounts';
-import Task from '../types/Task';
+import type { Task, TaskCounts } from "@/types";
+import { useEffect } from "react";
+import tableDataReducers from "../components/utils/tableDataReducers";
 
 interface useGetTaskCountsProps {
-  setTaskCounts: (taskCounts: TaskCounts) => void;
-  data: Task[] | undefined;
+	setTaskCounts: (taskCounts: TaskCounts) => void;
+	data: Task[] | undefined;
 }
 
 export default function useGetTaskCounts(props: useGetTaskCountsProps) {
-  const { setTaskCounts, data } = props;
+	const { setTaskCounts, data } = props;
 
-  useEffect(() => {
-    if (data) {
-      setTaskCounts({
-        all: tableDataReducers.allTasks(data).length,
-        today: tableDataReducers.today(data).length,
-        upcoming: tableDataReducers.upcoming(data).length,
-        overdue: tableDataReducers.overdue(data).length,
-        archived: tableDataReducers.archive(data).length,
-      });
-    }
-  }, [data, setTaskCounts]);
+	useEffect(() => {
+		if (data) {
+			setTaskCounts({
+				all: tableDataReducers.allTasks(data).length,
+				today: tableDataReducers.today(data).length,
+				upcoming: tableDataReducers.upcoming(data).length,
+				overdue: tableDataReducers.overdue(data).length,
+				archived: tableDataReducers.archive(data).length,
+			});
+		}
+	}, [data, setTaskCounts]);
 }
