+++ src/hooks/useGetTasks.ts
@@ -1,15 +1,15 @@
-import { useQuery } from '@tanstack/react-query';
-import axios from 'axios';
-import Task from '../types/Task';
-
-async function getTasks(): Promise<Task[]> {
-  const response = await axios.get('http://localhost:8000/tasks');
-  return response.data;
-}
+import api from "@/libs/api";
+import type { Task } from "@/types";
+import { useQuery } from "@tanstack/react-query";
 
 export default function useGetTasks() {
-  return useQuery({
-    queryKey: ['tasks'],
-    queryFn: getTasks,
-  });
+	async function getTasks(): Promise<Task[]> {
+		const response = await api.get("http://localhost:8000/tasks");
+		return response.data;
+	}
+
+	return useQuery({
+		queryKey: ["tasks"],
+		queryFn: getTasks,
+	});
 }
