+++ src/hooks/usePostTask.ts
@@ -1,18 +1,18 @@
-import { useMutation, useQueryClient } from '@tanstack/react-query';
-import axios from 'axios';
-import Task from '../types/Task';
+import api from "@/libs/api";
+import type { Task } from "@/types";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
 
 export default function usePostTask() {
-  const queryClient = useQueryClient();
+	const queryClient = useQueryClient();
 
-  const mutation = useMutation({
-    mutationFn: (task: Task) => {
-      return axios.post('http://localhost:8000/tasks', task);
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['tasks'] });
-    },
-  });
+	const mutation = useMutation({
+		mutationFn: (task: Task) => {
+			return api.post("http://localhost:8000/tasks", task);
+		},
+		onSuccess: () => {
+			queryClient.invalidateQueries({ queryKey: ["tasks"] });
+		},
+	});
 
-  return mutation;
+	return mutation;
 }
